requirements.txt
mysql-connector-python==8.0.27
python-telegram-bot==19.7
